feat: resolved comments and refactoring
